+++ CHANGELOG.md
@@ -1,7 +1,12 @@
+
+* NP-745 Move VR tests into /testing/visual-regression
+* NP-737 Install pa11y
+* NP-739 Create wcag-test task
+* NP-738 Create wcag-test:ci task
+
 ## <sub>v1.9.0</sub>
 
 #### _Sep. 7, 2020_
-
 * NP-744 Run Browser tests inside CI pipeline
 * NP-642 Header takes language switcher paramaters